+++ Nagru - Manga Organizer/Properties/AssemblyInfo.cs
@@ -34,5 +34,5 @@ using System.Resources;
 // by using the '*' as shown below:
 // [assembly: AssemblyVersion("1.0.*")]
 [assembly: AssemblyVersion("1.0.0.0")]
-[assembly: AssemblyFileVersion("4.6.28.0")]
+[assembly: AssemblyFileVersion("4.6.30.0")]
 [assembly: NeutralResourcesLanguageAttribute("en")]